opts now handles runtime arguments like I/O
kFastTracking takes one command line argument: the opts file.
The same change will come to the other programs soon.

The options file needs to be specified at runtime so we had to figure
out how to parse that argument.  Usually a options file based system is
much easier to maintain anyway (but one could also think of using
a robust opt=arg system like sqerp does).

The programs are not modular, so we may need a different template opts
file for each executable.  But they should probably all share the same
JobOptsSvc.  Would prefer to have on kTracker executable that takes the options file
as input.  Then the kTracker program would know which algorithms to call
and options to set based on the opts file.