add a threshold service singleton class
TresholdSvc allows you to set thresholds at
 None=0, Error, Warning, Info, Debug, Verbose=5
 <--------- increasing severity <----------
Asking the service Do<Level> will tell if if the severity if <LEVEL> is
equal to or greater than the threshold that is set.

The threshold is global and may be set using the job options file or in client code.

There are also shortcuts for each useful level comparison:
  error(), warning(), info(), debug(), verbose()
example:
  - set the threshold to Warning.
  - Threshold::info() returns false,  Threshold::error() returns true

All of this is in the Threshold namespace to protect enums and shortcut
functions.  You have three options
 - Use Threshold::<ThingInNamespace> everytime
 - Add "using namespace Threshold;" to everything in scope below has
   direct acccess to the namespace
 - Add "using Threshold::<level>;" to get access to just one element

The last method is probably the best.  For example, if you want access
to checks against live() add "using Threshold::live;"